Add loader.cancel() to cancel the loading thread
This function pushes a value to the is_done channel, which will
terminate the thread. Intended for use inside of love.quit() and only on love.quit().

Using it outside of a shutting down restart routine might lead to
unexpected behavior, since it is expected from you not needing those
resources anymore.